Move code out of BoogieGenerator.cs
