+++ Source/DafnyCore/Verifier/BoogieGenerator.Datatypes.cs
@@ -1,0 +1,139 @@
+﻿//-----------------------------------------------------------------------------
+//
+// Copyright (C) Microsoft Corporation.  All Rights Reserved.
+// Copyright by the contributors to the Dafny Project
+// SPDX-License-Identifier: MIT
+//
+//-----------------------------------------------------------------------------
+using System;
+using System.Collections.Generic;
+using System.Collections.Immutable;
+using System.Linq;
+using System.Numerics;
+using System.Diagnostics.Contracts;
+using Bpl = Microsoft.Boogie;
+using BplParser = Microsoft.Boogie.Parser;
+using System.Text;
+using System.Threading;
+using DafnyCore;
+using Microsoft.Boogie;
+using static Microsoft.Dafny.Util;
+using DafnyCore.Verifier;
+using JetBrains.Annotations;
+using Microsoft.Dafny.Triggers;
+using PODesc = Microsoft.Dafny.ProofObligationDescription;
+using static Microsoft.Dafny.GenericErrors;
+
+namespace Microsoft.Dafny;
+
+public partial class BoogieGenerator {
+  
+
+    /// <summary>
+    /// Return a sequence of expressions whose conjunction denotes a memberwise equality of "dt".  Recursive
+    /// codatatype equalities are written in one of the following ways:
+    /// If the codatatype equality is on a type outside the SCC of "dt", then resort to ordinary equality.
+    /// Else if the k==null, then:
+    ///   Depending on "limited", use the #2, #1, or #0 (limited) form of codatatype equality.
+    /// Else:
+    ///   Depending on "limited", use the #2, #1, or #0 (limited) form of prefix equality, passing "k"
+    ///   as the first argument.
+    /// </summary>
+    IEnumerable<Bpl.Expr> CoPrefixEquality(IToken tok, CoDatatypeDecl dt, List<Type> largs, List<Type> rargs, Bpl.Expr k, Bpl.Expr l, Bpl.Expr A, Bpl.Expr B, bool conjuncts = false) {
+      Contract.Requires(tok != null);
+      Contract.Requires(dt != null);
+      Contract.Requires(A != null);
+      Contract.Requires(B != null);
+      Contract.Requires(l != null);
+      Contract.Requires(Predef != null);
+      var etran = new ExpressionTranslator(this, Predef, dt.tok, dt);
+      // For example, for possibly infinite lists:
+      //     codatatype SList<T> = Nil | SCons(head: T, tail: SList<T>);
+      // produce with conjucts=false (default):
+      //   (A.Nil? && B.Nil?) ||
+      //   (A.Cons? && B.Cons? && A.head == B.head && Equal(k, A.tail, B.tail))
+      //
+      // with conjuncts=true:
+      //   (A.Nil? ==> B.Nil?) &&
+      //   (A.Cons? ==> (B.Cons? && A.head == B.head && Equal(k, A.tail, B.tail)))
+
+      Dictionary<TypeParameter, Type> lsu = Util.Dict(GetTypeParams(dt), largs);
+      Dictionary<TypeParameter, Type> rsu = Util.Dict(GetTypeParams(dt), rargs);
+
+      foreach (var ctor in dt.Ctors) {
+        Bpl.Expr aq = new Bpl.NAryExpr(tok, new Bpl.FunctionCall(GetReadonlyField(ctor.QueryField)), new List<Bpl.Expr> { A });
+        Bpl.Expr bq = new Bpl.NAryExpr(tok, new Bpl.FunctionCall(GetReadonlyField(ctor.QueryField)), new List<Bpl.Expr> { B });
+        Bpl.Expr chunk = Bpl.Expr.True;
+        foreach (var dtor in ctor.Destructors) {  // note, ctor.Destructors has a field for every constructor parameter, whether or not the parameter was named in the source
+          var a = new Bpl.NAryExpr(tok, new Bpl.FunctionCall(GetReadonlyField(dtor)), new List<Bpl.Expr> { A });
+          var b = new Bpl.NAryExpr(tok, new Bpl.FunctionCall(GetReadonlyField(dtor)), new List<Bpl.Expr> { B });
+          var ty = dtor.Type;
+          Bpl.Expr q;
+          var codecl = ty.AsCoDatatype;
+          if (codecl != null && codecl.SscRepr == dt.SscRepr) {
+            var lexprs = Map(ty.TypeArgs, tt => tt.Subst(lsu));
+            var rexprs = Map(ty.TypeArgs, tt => tt.Subst(rsu));
+            q = CoEqualCall(codecl, lexprs, rexprs, k, l, a, b);
+          } else {
+            // ordinary equality; let the usual translation machinery figure out the translation
+            var tyA = ty.Subst(lsu);
+            var tyB = ty.Subst(rsu);
+            var aa = CondApplyUnbox(tok, a, ty, tyA);
+            var bb = CondApplyUnbox(tok, b, ty, tyB);
+            var equal = new BinaryExpr(tok, BinaryExpr.Opcode.Eq, new BoogieWrapper(aa, tyA), new BoogieWrapper(bb, tyB));
+            equal.ResolvedOp = ModuleResolver.ResolveOp(equal.Op, tyA, tyB);  // resolve here
+            equal.Type = Type.Bool;  // resolve here
+            q = etran.TrExpr(equal);
+          }
+          chunk = BplAnd(chunk, q);
+        }
+        if (conjuncts) {
+          yield return Bpl.Expr.Binary(new NestedToken(tok, ctor.tok), BinaryOperator.Opcode.Imp, aq, BplAnd(bq, chunk));
+        } else {
+          yield return BplAnd(BplAnd(aq, bq), BplImp(BplAnd(aq, bq), chunk));
+        }
+      }
+    }
+    
+
+    // Makes a call to equality, if k is null, or otherwise prefix equality. For codatatypes.
+    Bpl.Expr CoEqualCall(CoDatatypeDecl codecl, List<Bpl.Expr> largs, List<Bpl.Expr> rargs, Bpl.Expr k, Bpl.Expr l, Bpl.Expr A, Bpl.Expr B, Bpl.IToken tok = null) {
+      Contract.Requires(codecl != null);
+      Contract.Requires(largs != null);
+      Contract.Requires(rargs != null);
+      Contract.Requires(l != null);
+      Contract.Requires(A != null);
+      Contract.Requires(B != null);
+      if (tok == null) {
+        tok = A.tok;
+      }
+      List<Bpl.Expr> args = Concat(largs, rargs);
+      if (k != null) {
+        args.Add(k);
+      }
+      args.AddRange(new List<Bpl.Expr> { l, A, B });
+      var fn = k == null ? CoEqualName(codecl) : CoPrefixName(codecl);
+      return FunctionCall(tok, fn, Bpl.Type.Bool, args);
+    }
+
+    // Same as above, but with Dafny-typed type-argument lists
+    Bpl.Expr CoEqualCall(CoDatatypeDecl codecl, List<Type> largs, List<Type> rargs, Bpl.Expr k, Bpl.Expr l, Bpl.Expr A, Bpl.Expr B, IToken tok = null) {
+      Contract.Requires(codecl != null);
+      Contract.Requires(largs != null);
+      Contract.Requires(rargs != null);
+      Contract.Requires(l != null);
+      Contract.Requires(A != null);
+      Contract.Requires(B != null);
+      return CoEqualCall(codecl, Map(largs, TypeToTy), Map(rargs, TypeToTy), k, l, A, B, tok);
+    }
+
+    static string CoEqualName(CoDatatypeDecl codecl) {
+      Contract.Requires(codecl != null);
+      return "$Eq#" + codecl.FullSanitizedName;
+    }
+
+    static string CoPrefixName(CoDatatypeDecl codecl) {
+      Contract.Requires(codecl != null);
+      return "$PrefixEq#" + codecl.FullSanitizedName;
+    }
+}
